Adding random suffix to apple appstore unclaimed accounts
There could be a scenario, where VerifyPurchase would attempt
to create an account that already existed which caused an error:

- Verify purchase for anonymous device token.
- Create original transaction ID based account.
- Claim account (old account is only disabled, email is still taken).
- Reset device token on the device.
- Verify purchase for anonymous device token.

Attempt to create original transaction ID based account would fail here
because account with that original transaction ID already exists.
As it's not safe to reinstantiate already disabled accounts, it's
better to add suffixes and avoid conflicts completely.