Fix duplicate prepaid payments for Apple Appstore
If charge was executed before the end of current subscription,
the condition was not evaluated correctly. This state caused
the charge to be "successful" eventhough Apple hasn't charged
anything yet:

  $subscriptionEndDate: 2020-12-30 15:00 (set based on the receipt
    expiration found after the last successful payment)

  $receiptExpiration: 2020-12-30 15:00 (found in this attempt to
    charge the payment)

The $receiptExpiration clearly says, that the subscription
was not yet prolonged, but the condition allowed it to continue
and process the payment as if it was already charged/prolonged.

If the receiptExpiration equals the end date of currently active
subscription, it should be treated as a fail try.

remp/crm#1566