Fix unnecessary warning regarding `exclude_old_transactions`
This option in Apple's validation request doesn't guarantee only one
receipt is returned in the response. It guarantees that only "one
transaction for each susbcription" is returned.

If user has multiple subscriptions (one older, one newer), multiple
receipts are a valid state.

Because of that:

- If we already know original transaction ID, we filter the receipt
based on that.
- If we don't know it, we use the first receipt in the array (should
be the latest).

remp/crm#2515