Fix possibly discarded server-to-server notification
The error ocurred if both cancellation and subscription fields are present.

Original schema expected, that either cancellation OR subscriptions fields
are present, but not all of them.

The downside of the current schema is, that if subscription fields are present
and only one of two cancellation fields is present, the schema is still considered
valid.

Possibility of missing cancel reason was already handler in
https://github.com/remp2020/crm-apple-appstore-module/commit/88293f9f323927b5b62a917420d1c15aba049c61

remp/crm#2516