Added handling for CONSUMPTION_REQUEST and REFUND_DECLINED
We decided not to actively handle CONSUMPTION_REQUEST
notifications because of the requirement to get user's
consent we can share the data with Apple within 12 hours
of this request.

The REFUND_DECLINED is only informative notification
and since we don't initiate refunds ourselves, we ignore
this notification and expect the initiating party
to handle this on their own.

remp/crm#3162