Fix inconsistency of inapp sendEmail flags
- The initial payment (confirmed via VerifyPurchase API) was
always emitted with sendEmail: false flag, causing this payment
never to appear in the scenarios.
- The renewal payment (confirmed via webhook) was always
emitted with sendEmail: true flag, because it was processed
through RecurrentPaymentProcessor.
- The renewal payment (confirmed via gateway, used as fallback)
was always emitted with sendEmail: false. This however isn't
the primary implementation and a small number of payments
are actually processed here.

To make this consistent, we now set the sendEmail flag always
to true, so that your scenarios can handle these kind of payments
and subscriptions if they need to, for example for onboarding
emails.

Your scenarios can still filter these subscriptions out either
by using "subscription type" condition, or "has enabled notifications"
condition.

remp/novydenik#1385